Add csv parser
The short story is: the department of buildings stopped putting the
excel files on their website. However, they do seen to update nyc open data with
some regularity.

Therefore the 'csvparser' will now update the database with that
data. I also changed the schema slightly, so the rest of the app will
need to be adjusted (and this  will eventually make the old parser obsolete).